ai-expert-simulator/backend/requirements.txt
Flask
python-dotenv
Flask-CORS
openai
# Add other dependencies like PyMongo or MongoEngine later 